refactor: folder workflow
