+++ app/dashboard/project/[id]/workflows/steps/shared/components/DocumentPreviewModal.tsx
