+++ app/dashboard/project/[id]/workflows/steps/shared/components/GeneratedDocumentsList.tsx
