+++ app/dashboard/project/[id]/workflows/steps/shared/hooks/useDocumentPreview.ts
