+++ package.json
@@ -30,7 +30,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "diff": "^8.0.3",
-    "github-markdown-css": "^5.8.1",
+    "github-markdown-css": "^5.9.0",
     "input-otp": "^1.4.2",
     "lucide-react": "^0.544.0",
     "mermaid": "^11.12.1",
